# swagger/swagger.yaml
swagger: "2.0"
info:
  description: "hello world"
  version: "1.0.0"
  title: "my api-doc"
host: "localhost:9000"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /api/buy/countAuction:
    get:
      tags:
      - "buy"
      summary: "count auction number"
      description: ""
      operationId: "countAuction"
      produces:
      - "application/json"
      parameters:
      - name: "user_id"
        in: "query"
        description: "지금은 그냥 user_id = 1 로 생각하셈"
        required: false
        type: "string"
        collectionFormat: "multi"
      responses:
        "200":
          description: "successful operation"
        "101":
          description: "Invalid status value"
  /api/buy/getStep1Latest:
    get:
      tags:
      - "buy"
      summary: "get 6 phones at /web/buy/1"
      description: "if you already did a selection, you also get the selected output"
      operationId: "getStep1Latest"
      produces:
      - "application/json"
      parameters:
      - name: "user_id"
        in: "query"
        description: "지금은 그냥 user_id = 1 로 생각하셈"
        required: false
        type: "string"
        collectionFormat: "multi"
      responses:
        "200":
          description: "successful operation"
        "101":
          description: "Invalid status value"
  /api/buy/getStep1WithBrand:
    get:
      tags:
      - "buy"
      summary: "Finds devices with brand_id"
      description: "at buy/1/brand"
      operationId: "getStep1WithBrand"
      produces:
      - "application/json"
      parameters:
      - name: "brand_id"
        in: "query"
        description: "1: Samsung, 2: Apple, 3:LG, 4:Others"
        required: true
        type: "integer"
        collectionFormat: "multi"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "Invalid status value"
  /api/buy/postSaveStep1:
    post:
      tags:
      - "buy"
      summary: "save data from step 1"
      description: "step=1, state=1, update(or insert) user_id, device_id"
      operationId: "postSaveStep1"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "user_id"
        in: "query"
        description: "give current user_id"
        required: true
        type: "string"
        collectionFormat: "multi"
      - name: "device_id"
        in: "query"
        description: "give the selected device's id"
        required: true
        type: "integer"
        collectionFormat: "multi"
      responses:
        "405":
          description: "Invalid input"
  /api/buy/getStep2ColorVolume:
    get:
      tags:
      - "buy"
      summary: "Finds device color and volume with user_id"
      description: "at buy/2"
      operationId: "getStep2ColorVolume"
      produces:
      - "application/json"
      parameters:
      - name: "user_id"
        in: "query"
        description: "agency(1:SKT ,2:KT, 3:LG, 4:알뜰) volume은 그냥 int형으로!\n 지금은 device_id = 1,2만 다르고 다른건 다 device_id=1꺼 복붙해놓음"
        required: true
        type: "string"
        collectionFormat: "multi"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "Invalid status value"
  /api/buy/postSaveStep2:
    post:
      tags:
      - "buy"
      summary: "save data from step 2"
      description: "Update device_detail_id(info about color, volume etc)"
      operationId: "postSaveStep2"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "user_id"
        in: "query"
        description: "give current user_id"
        required: true
        type: "string"
        collectionFormat: "multi"
      - name: "device_detail_id"
        in: "query"
        description: "give the selected device_detail's id"
        required: true
        type: "integer"
        collectionFormat: "multi"
      responses:
        "405":
          description: "Invalid input"
  /api/buy/getStep3PaymentInfo:
    get:
      tags:
      - "buy"
      summary: "요금제 정보 가져오기"
      description: "agency(1:SKT ,2:KT, 3:LG), generation(4: 4G, 5:5G)"
      operationId: "getStep3PaymentInfo"
      produces:
      - "application/json"
      parameters:
      - name: "agency"
        in: "query"
        description: ""
        required: true
        type: "integer"
        collectionFormat: "multi"
      - name: "generation"
        in: "query"
        description: ""
        required: true
        type: "integer"
        collectionFormat: "multi"
      responses:
        "200":
          description: ""
        "400":
          description: "Invalid status value"
  /api/buy/getStep3OfficialInfo:
    get:
      tags:
      - "buy"
      summary: "공시지원금 정보 가져오기"
      description: "returns discount_official"
      operationId: "getStep3OfficialInfo"
      produces:
      - "application/json"
      parameters:
      - name: "payment_id"
        in: "query"
        description: ""
        required: true
        type: "integer"
        collectionFormat: "multi"
      - name: "device_detail_id"
        in: "query"
        description: ""
        required: true
        type: "integer"
        collectionFormat: "multi"
      responses:
        "200":
          description: ""
        "400":
          description: "Invalid status value"
  /api/buy/getStep3Info:
    get:
      tags:
      - "buy"
      summary: "Finds selected device and color contraint(agency) with user_id"
      description: ""
      operationId: "getStep3Info"
      produces:
      - "application/json"
      parameters:
      - name: "user_id"
        in: "query"
        description: ""
        required: true
        type: "string"
        collectionFormat: "multi"
      responses:
        "200":
          description: "agency(1:SKT ,2:KT, 3:LG, 4:알뜰)"
        "400":
          description: "Invalid status value"
  /api/buy/postSaveStep3:
    post:
      tags:
      - "buy"
      summary: "save data from step 3"
      description: "Update auction(with info on step3 and device_detail_id, user_id)"
      operationId: "postSaveStep3"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "user_id"
        in: "query"
        description: "give current user_id"
        required: true
        type: "string"
        collectionFormat: "multi"
      - name: "agency_use"
        in: "query"
        description: "현재 통신사"
        required: true
        type: "integer"
        collectionFormat: "multi"
      - name: "agency_hope"
        in: "query"
        description: "사용희망 통신사"
        required: true
        type: "integer"
        collectionFormat: "multi"
      - name: "payment_id"
        in: "query"
        description: "요금제 id"
        required: true
        type: "integer"
        collectionFormat: "multi"
      - name: "period"
        in: "query"
        description: "할부기간"
        required: true
        type: "integer"
        collectionFormat: "multi"
      - name: "contract_list"
        in: "query"
        description: "고급옵션: '1,1,2,2,1,2,1' 이런식으로 줄것"
        required: true
        type: "string"
        collectionFormat: "multi"
      responses:
        "405":
          description: "Invalid input"
...
  /api/myAuction/get201MyAuctionOn:
    get:
      tags:
      - "myAuction"
      summary: "나의경매-입찰중 page"
      description: "203 or 204 페이지로 갈때 auction_id 넘겨야함"
      operationId: "get201MyAuctionOn"
      produces:
      - "application/json"
      parameters:
      - name: "user_id"
        in: "query"
        description: ""
        required: true
        type: "string"
        collectionFormat: "multi"
      responses:
        "200":
          description: ""
        "400":
          description: "Invalid status value"
  /api/myAuction/get201StateUpdate:
    post:
      tags:
      - "myAuction"
      summary: "나의경매-입찰중 page"
      description: "203 or 204 페이지로 갈때 finish_time 서버에서 비교해서 state값 넘겨줌"
      operationId: "get201StateUpdate"
      produces:
      - "application/json"
      parameters:
      - name: "finish_time"
        in: "query"
        description: ""
        required: true
        type: "string"
        collectionFormat: "multi"
      responses:
        "200":
          description: ""
        "400":
          description: "Invalid status value"
  /api/myAuction/get202MyAuctionOff:
    get:
      tags:
      - "myAuction"
      summary: "지난경매 page"
      description: "205 페이지로 갈때 front 에서 win_deal_id 넘겨야함"
      operationId: "get202MyAuctionOff"
      produces:
      - "application/json"
      parameters:
      - name: "user_id"
        in: "query"
        description: ""
        required: true
        type: "string"
        collectionFormat: "multi"
      responses:
        "200":
          description: ""
        "400":
          description: "Invalid status value"
  /api/myAuction/get203MyAuctionDetails:
    get:
      tags:
      - "myAuction"
      summary: "나의경매-상세 마감전"
      description: "211 (store상세)api 갈때 store_id 넘겨주기. \n 206 페이지로 갈때 front 에서 deal_id 넘겨야함."
      operationId: "get203MyAuctionDetails"
      produces:
      - "application/json"
      parameters:
      - name: "auction_id"
        in: "query"
        description: ""
        required: true
        type: "integer"
        collectionFormat: "multi"
      responses:
        "200":
          description: ""
        "400":
          description: "Invalid status value"
  /api/myAuction/get204MyAuctionDetailsFinish:
    get:
      tags:
      - "myAuction"
      summary: "나의경매 상세 마감&낙찰 직전"
      description: "211 (store상세)api로 store_id넘겨주기.\n206 페이지로 갈때 front 에서 deal_id 넘겨야함."
      operationId: "get204MyAuctionDetailsFinish"
      produces:
      - "application/json"
      parameters:
      - name: "auction_id"
        in: "query"
        description: ""
        required: true
        type: "integer"
        collectionFormat: "multi"
      responses:
        "200":
          description: ""
        "400":
          description: "Invalid status value"
  /api/myAuction/get205DealDetails:
    get:
      tags:
      - "myAuction"
      summary: "나의경매-상세 낙찰 후"
      description: "205, 206, 207page 모두 경메 상세내용 볼때 이거사용\n205->210(리뷰 페이지) 갈때 front 에서 deal_id 넘겨야함.\n206->208(경매낙찰 팝업) 뜰때 front에서 (deal_id, store_name) 넘김\n207->208(경매낙찰 팝업) 뜰때 front에서 (deal_id, store_name) 넘김"
      operationId: "get205DealDetails"
      produces:
      - "application/json"
      parameters:
      - name: "deal_id"
        in: "query"
        description: ""
        required: true
        type: "integer"
        collectionFormat: "multi"
      responses:
        "200":
          description: ""
        "400":
          description: "Invalid status value"
  /api/myAuction/patch208ConfirmPopup:
    patch:
      tags:
      - "myAuction"
      summary: "낙찰 페이지"
      description: "209로 갈때  deal_id 넘김"
      operationId: "patch208ConfirmPopup"
      produces:
      - "application/json"
      parameters:
      - name: "deal_id"
        in: "query"
        description: ""
        required: true
        type: "integer"
        collectionFormat: "multi"
      responses:
        "200":
          description: ""
        "400":
          description: "Invalid status value"
  /api/myAuction/get209ConfirmedAuction:
    get:
      tags:
      - "myAuction"
      summary: "경매 낙찰 완료"
      description: ""
      operationId: "get209ConfirmedAuction"
      produces:
      - "application/json"
      parameters:
      - name: "deal_id"
        in: "query"
        description: ""
        required: true
        type: "integer"
        collectionFormat: "multi"
      responses:
        "200":
          description: ""
        "400":
          description: "Invalid status value"
  /api/myAuction/get210InfoForReview:
    get:
      tags:
      - "myAuction"
      summary: "리뷰 작성-필요 정보"
      description: ""
      operationId: "get210InfoForReview"
      produces:
      - "application/json"
      parameters:
      - name: "deal_id"
        in: "query"
        description: ""
        required: true
        type: "integer"
        collectionFormat: "multi"
      responses:
        "200":
          description: ""
        "400":
          description: "Invalid status value"
  /api/myAuction/post210DealReview:
    post:
      tags:
      - "myAuction"
      summary: "리뷰 작성-입력하기"
      description: "210 -> 211(store상세) 갈때 store_id 넘김"
      operationId: "post210DealReview"
      produces:
      - "application/json"
      parameters:
      - name: "score"
        in: "query"
        description: ""
        required: true
        type: "integer"
        collectionFormat: "multi"
      - name: "comment"
        in: "query"
        description: ""
        required: true
        type: "string"
        collectionFormat: "multi"
      - name: "deal_id"
        in: "query"
        description: ""
        required: true
        type: "integer"
        collectionFormat: "multi"
      responses:
        "200":
          description: ""
        "400":
          description: "Invalid status value"
  /api/myAuction/get211StoreDetails:
    get:
      tags:
      - "myAuction"
      summary: " store 상세"
      description: "212로 갈때 store_id 넘겨주기"
      operationId: "get211StoreDetails"
      produces:
      - "application/json"
      parameters:
      - name: "deal_id"
        in: "query"
        description: ""
        required: true
        type: "integer"
        collectionFormat: "multi"
      responses:
        "200":
          description: ""
        "400":
          description: "Invalid status value"  
  /api/myAuction/get212AllStoreReviews:
    get:
      tags:
      - "myAuction"
      summary: "리뷰 리스트"
      description: ""
      operationId: "get212AllStoreReviews"
      produces:
      - "application/json"
      parameters:
      - name: "store_id"
        in: "query"
        description: ""
        required: true
        type: "string"
        collectionFormat: "multi"
      responses:
        "200":
          description: ""
        "400":
          description: "Invalid status value"
...